backend/requirements.txt
# ─── Web Server & API ────────────────────────────────────────────────────────
fastapi==0.136.1
uvicorn==0.46.0
gunicorn==25.3.0
python-multipart==0.0.27
email-validator==2.3.0
slowapi==0.1.9

# ─── Database & Auth ────────────────────────────────────────────────────────
supabase==2.29.0
python-dotenv==1.0.1
python-jose==3.5.0
passlib==1.7.4
PyJWT==2.12.1
bcrypt==5.0.0
cryptography==47.0.0

# ─── AI & LLMs ──────────────────────────────────────────────────────────────
google-generativeai==0.8.6
google-genai==1.74.0
openai==2.30.0
litellm==1.83.7

# ─── Payments & Integrations ───────────────────────────────────────────────
stripe==15.1.0
razorpay==2.0.1

# ─── Utils & Performance ───────────────────────────────────────────────────
pydantic==2.12.5
pydantic-settings==2.14.0
httpx==0.28.1
requests==2.33.1
python-dateutil==2.9.0.post0
tenacity==9.1.4
pillow==12.2.0
pandas==3.0.2
numpy==2.4.4
tqdm==4.67.3

# ─── Observability ─────────────────────────────────────────────────────────
sentry-sdk==2.58.0
python-json-logger==4.1.0
